+++ src/executable.cpp
@@ -1,8 +1,11 @@
 #include "executable.hpp"
-
+#include <sys/types.h>
+#include <sys/stat.h>
+#include <fcntl.h>
 #include <unistd.h>
 #include <cstdlib>
 #include <stdexcept>
+#include <cerrno>
 #include <sstream>
 
 #include "configuration.hpp"
@@ -37,7 +40,6 @@ parse_configuration
             case 'v':
                 parsed_configuration.verbose_ = true;
                 break;
-                break;
             case 'h':
                 parsed_configuration.print_help_ = true;
                 break;
@@ -62,7 +64,33 @@ void
 daemonize
     ( configuration const& parsed_configuration )
 {
-    throw because() << "daemonizing is unimplemented";
+    pid_t const pid = ::fork();
+    if ( pid < 0)
+        throw because() << "can't fork '" << ::strerror( errno ) << "'";
+
+    // If this is parent process, exit.
+    if ( pid )
+        return;
+
+    ::umask( 0 );
+
+    // Detach from parent session.
+    if ( ::setsid() < 0 )
+        throw because() << "can't create a new process session '" 
+                << ::strerror( errno ) << "'";
+
+    if ( ::chdir( "/" ) < 0 )
+        throw because() << "can't chdir to / '" << ::strerror( errno ) << "'";
+    
+    if ( ::close( STDIN_FILENO ) < 0 )
+        throw because() << "can't close stdin '" << ::strerror( errno ) << "'";
+
+    if ( ::open ("/dev/null", O_RDWR) != STDIN_FILENO )
+        throw because() << "can't open /dev/null '" << ::strerror( errno ) << "'";
+        
+    if ( ::dup2( 0, STDOUT_FILENO ) < 0 || ::dup2( 0, STDERR_FILENO ) < 0 )
+        throw because() << "can't reassign stdout & stderr'" 
+                << ::strerror( errno ) << "'";
 
     main_loop::run( parsed_configuration );
 }